pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "channels>=4.2.2",
    "channels-redis>=4.2.1",
    "cryptography>=45.0.4",
    "dj-database-url>=3.0.0",
    "django>=5.2.3",
    "django-allauth>=65.9.0",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-prometheus>=2.4.0",
    "django-redis>=6.0.0",
    "django-split-settings>=1.3.2",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "pillow>=11.3.0",
    "psycopg>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
    "uv>=0.7.13",
    "uvicorn>=0.34.3",
    "uwsgi>=2.0.30",
]

[tool.setuptools.packages.find]
include = ["config*", "oauth*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = ["test_*.py", "*_test.py", "tests.py"]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-django",
    "pytest-cov",
]
